Line 163 removed "ERROR" for mounted drives
I use an email rule in my Gmail to filter server emails to a folder/label and mark as read, with the exception if the subject or body contains the word ERROR.
It is annoying you get ERROR emails just because a backup drive was already mounted (I mounted it manually to restore something and forgot to unmount it). Now, the error is not triggered for this device.
I also made sure it is added to the list of mounted drives, to unsure the script will unmount the drive when done. Previously, the script did not unmount already mounted drives, only drives that were mounted by the script.